refacor: create scene manager and move start scene
